Changes made for Week7 Day 4.
