--- data/courier_list.csv
@@ -1,7 +1,0 @@
-Id,Name,Age,Vehicle,Status
-1,Martin,27,Car,Busy
-2,Sheila,41,Car,Busy
-3,Mac,50,Car,Free
-4,Ozer,31,Bike,Free
-5,Wei,21,Bike,Busy
-6,Lilly,18,Bike,Free
