--- data/drinks_list.csv
@@ -1,14 +1,0 @@
-Id,Drink,Type,Price,Status
-1,Tea,Hot,£1.25,Available
-2,Coffee,Hot,£1.50,Available
-3,Decaf Coffe,Hot,£1.50,Out Of Stock
-4,Hot Chocolate,Hot,£1.80,Available
-5,Fanta,Cold,£1.00,Available
-6,Sprite,Cold,£1.00,Available
-7,Coca Cola,Cold,£1.00,Available
-8,Coke Zero,Cold,£1.25,Out Of Stock
-9,Water,Cold,£0.50,Available
-10,Beer,Alcoholic,£3.00,Available
-11,Red Wine,Alcoholic,£12,Available
-12,White Wine,Alcoholic,£11,Out Of Stock
-13,Spirits,Alcoholic,£20.00,Available
