+++ rating_funcs.py
@@ -1,3 +1,10 @@
+#%%
+import matplotlib.pyplot as plt
+import matplotlib as mpl
+import numpy as np
+
+
+
 
 def rate_sys(connection):
     rate = input(
@@ -16,43 +23,28 @@ def rate_sys(connection):
             f'INSERT INTO ratings (username, rating) VALUES ("{rate_name}", {rate})'
         )
         connection.commit()
-        cursor.close()
         print("Thanks, we'll take that onboard. Until next time!")
+        cursor.execute('select * from ratings')
+        ratingtable = cursor.fetchall()
+        username = []
+        rating = []
+        for row in ratingtable :
+            username.append(row[1])
+            rating.append(row[2])
+
+
+        x_pos = np.arange(len(username))
+        plt.bar(x_pos, rating)
+        plt.xticks(x_pos, username)
+        plt.ylabel('Rating')
+        plt.xlabel('Usernames')
+        plt.title('User Ratings')
+        plt.show()
     elif rate in rate_n:
         print("Ok no problem. Until next time!")
     else:
         print("Sorry, invalid selection. Please try again.")
     return rate
 
-#%%
-import matplotlib.pyplot as plt
-import matplotlib as mpl
-import numpy as np
-
-%matplotlib inline
-
-%load_ext sql
-
-%sql mysql+pymysql://root:password@localhost:3306/mini_project
-
-ratingtable = %sql select * from ratings
-username = []
-rating = []
-for row in ratingtable :
-    username.append(row[1])
-    rating.append(row[2])
-
-
-x_pos = np.arange(len(username))
-plt.bar(x_pos, rating)
-plt.xticks(x_pos, username)
-plt.ylabel('Rating')
-plt.xlabel('Usernames')
-plt.title('User Ratings')
-plt.show()
-
-
-
-# %%
 
 # %%
